requirements.txt
fastapi
uvicorn
slack_sdk
google-api-python-client
python-dotenv
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib